# API_CONTRACTS.yaml
...
openapi: 3.0.3
info:
  title: QADash Enterprise API
  description: |
    Enterprise AI-Driven Testing Platform API
    
    ## Authentication
    All endpoints require authentication via JWT Bearer token.
    Include the token in Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Default: 100 requests/minute
    - Authenticated: 1000 requests/minute
    - Enterprise: Unlimited
    
    ## Response Format
    All responses follow standard JSON format with success status.
  version: 1.0.0
  contact:
    name: QADash Support
    email: support@qadash.io
    url: https://qadash.io/support
  license:
    name: Enterprise License
    url: https://qadash.io/license
servers:
  - url: https://api.qadash.io/v1
    description: Production Server
  - url: https://staging-api.qadash.io/v1
    description: Staging Server
  - url: https://dev-api.qadash.io/v1
    description: Development Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /executions:
    post:
      tags: [Execution]
      summary: Create execution
      operationId: createExecution
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExecutionInput'
      responses:
        '201':
          description: Execution created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'
    
    get:
      tags: [Execution]
      summary: List executions
      operationId: listExecutions
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionListResponse'
  /executions/{id}:
    parameters:
      - $ref: '#/components/parameters/executionId'
    
    get:
      tags: [Execution]
      summary: Get execution details
      operationId: getExecution
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionDetail'
    
    delete:
      tags: [Execution]
      summary: Cancel execution
      operationId: cancelExecution
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Execution cancelled
  /executions/{id}/retry:
    parameters:
      - $ref: '#/components/parameters/executionId'
    
    post:
      tags: [Execution]
      summary: Retry execution
      operationId: retryExecution
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                failedOnly:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Retry initiated

  # ==========================================
  # AI ENDPOINTS
  # ==========================================
...
  /bugs:
    post:
      tags: [Bugs]
      summary: Create bug
      operationId: createBug
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBugInput'
      responses:
        '201':
          description: Bug created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
    
    get:
      tags: [Bugs]
      summary: List bugs
      operationId: listBugs
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: List of bugs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BugListResponse'
  /bugs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    
    get:
      tags: [Bugs]
      summary: Get bug details
      operationId: getBug
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bug details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BugDetail'
    
    patch:
      tags: [Bugs]
      summary: Update bug
      operationId: updateBug
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBugInput'
      responses:
        '200':
          description: Bug updated

  # ==========================================
  # MOBILE ENDPOINTS
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  parameters:
...
    executionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique execution identifier
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items per page
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Orchestration schemas
...
    CreateExecutionInput:
      type: object
      required: [name, projectId, testIds]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        projectId:
          type: string
          format: uuid
        testIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        environmentId:
          type: string
          format: uuid
        browser:
          type: string
          enum: [chromium, firefox, webkit]
          default: chromium
        options:
          type: object
          properties:
            parallel:
              type: boolean
            sharding:
              type: integer
            viewport:
              type: object
    Execution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        projectId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in milliseconds
    ExecutionDetail:
      allOf:
        - $ref: '#/components/schemas/Execution'
        - type: object
          properties:
            summary:
              type: object
              properties:
                total:
                  type: integer
                passed:
                  type: integer
                failed:
                  type: integer
                skipped:
                  type: integer
                passRate:
                  type: number
            testResults:
              type: array
              items:
                $ref: '#/components/schemas/TestResult'
            logs:
              type: string
              description: Execution logs
    ExecutionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Execution'
        pagination:
          $ref: '#/components/schemas/Pagination'
    TestResult:
      type: object
      properties:
        testId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [passed, failed, skipped, pending]
        duration:
          type: integer
        error:
          type: string
        screenshots:
          type: array
          items:
            type: string
        videoUrl:
          type: string

    # AI schemas
...
    CreateBugInput:
      type: object
      required: [title, severity]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assigneeId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
    UpdateBugInput:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        assigneeId:
          type: string
          format: uuid
        resolution:
          type: string
          enum: [fixed, wont_fix, duplicate, cant_fix]
    Bug:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BugDetail:
      allOf:
        - $ref: '#/components/schemas/Bug'
        - type: object
          properties:
            assignee:
              type: object
            reporter:
              type: object
            linkedExecutions:
              type: array
              items:
                type: string
                format: uuid
            comments:
              type: array
              items:
                type: object
    BugListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Bug'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Mobile schemas
...